refactor helper decorator and tests all to helpers.py, test_helpers.py
